Modularize domain layer
