+++ app/build.gradle.kts
@@ -75,6 +75,7 @@ ktlint {
 }
 
 dependencies {
+    implementation(project(":domain"))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
