+++ build.gradle.kts
@@ -4,4 +4,5 @@ plugins {
     alias(libs.plugins.kotlin.android) apply false
     id("com.google.dagger.hilt.android") version "2.51.1" apply false
     id("org.jlleitschuh.gradle.ktlint") version "11.3.1" apply false
+    alias(libs.plugins.android.library) apply false
 }
