+++ domain/src/main/java/com/hpgallery/domain/repository/HpPreferencesRepository.kt
