+++ gradle/libs.versions.toml
@@ -36,6 +36,7 @@ timber = "5.0.1"
 uiTestJunit4 = "1.4.0"
 uiTooling = "1.5.0"
 uiToolingPreview = "1.5.0"
+android-gradlePlugin = "8.6.0"
 
 [libraries]
 androidx-core = { module = "androidx.test:core", version.ref = "core" }
@@ -89,4 +90,5 @@ ui-tooling-preview = { module = "androidx.compose.ui:ui-tooling-preview", versio
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
+android-library = { id = "com.android.library", version.ref = "android-gradlePlugin" }
